"Fix" order of files in advanced search hierarchy
"Fix" in quotes here because this solution is far from ideal, but I
could not think of a better way to do this at the moment. Python does
not seem to have a builtin capable of sorting strings in the same way
Postgres does by default, so we sort the files in Postgres. I left some
more comments on this in the code.