Refactor network trace cache to use data hash
Previously, sankey network traces were stored in sessionStorage, where
they were latered retrieved using a unique ID. This was limited, since a
user could not open the trace in a new tab, or share the trace with
other users (or even paste the link into a seperate browser window).

Instead, I've introduced (more or less) permanent links for the traces.
This is done by hashing a set of (again, more or less) unique data for
the trace, and using both that hash and the sankey file id as the URL
for the trace.