Remove organism from find synonym submission
Previously were adding the organism matched to the selected synonyms
when submitting the "Find Synonyms" dialog. However, this was a
restrictive default, so we decided to remove the organism in favor of
requiring that the user add it later if they are truly interested in
documents that must contain the organism.