Fix misc. bugs in content search
Refactors how the query string is managed in the component and its
children. Also fixes a bug where there was whitespace between the query
and synonyms when they are added. Finally, fixes a bug where selecting a
folder and closing the advanced search component was not consistent with
other fields.