Fix project cloud search + content search form
Fixes an issue where the content search endpoint wasn't properly used.
Looks like the erroneous code was using an older set of params.

Also fixes an issue in the content search form where the user could not
submit by pressing "Enter". This was an oversight introduced in a recent
refactoring of the form.