Prioritize matches where primary name is the same as synonym.
Do this check during the LMDB lookups. This seems like the most reasonable place to do this, because we can
potentially filter out lots of unwanted matches. If there is no match from the filter, then use the unfiltered results from LMDB.

Also changed a bit how the gene/protein organism matching is done. Instead of adding all possible results from LMDB,
filter out using a set. Since we only care about the synonyms, we can potentially reduce a significant number of
items in the list with a set. We don't care about the primary name, since we can update that with the entity id.

Also filtered out the results from the graph as well, prioritizing primary names that are the same as their synonym.
This will help make the code a bit cleaner and not have to use next on a dict.