New CloudStorage classes and new create_data_file.py.
The create_data_file.py is used (as a starting point) to create graph .tsv data files that will be used alongside migrations.
This will help us keep track and know what is in the graph database. The script queries and creates the .tsv file, then
uploads that to Azure. Migration will then pull that file and run against the graph (still TODO).

Removed the  folder as that was used to load global inclusions, but the script is now very outdated and no longer needed.