Add uniqueness criteria to adv search type dropdwn
Because both many-to-many sankey and the "regular" sankey module provide
the graph type provider, it appears twice in the list when we get them
all. As a temporary fix, we filter non-unique values from the provider
list in the advanced content search dropdown. This "fixes" the bug, but
isn't meant to be a permanent solution. Either we need to commit to the
two views using a different file type provider, or we need to redesign
the split such that they both use the same provider, and the provider is
only provided once.