Clean up search logic in visjs viewer
Two things: 1) there was actually a bug in the template, we were calling
the `findNext` behavior an additional time. Getting rid of the
`enterPress` callback resolved this. 2) The component has also been
updated to reflect this fix. Previously we were setting the initial
search index to -1, but this was really a symptom of the problem in #1.