Added StringDB liquibase and fixed Zenodo liquibase.
The previous cypher queries for Zenodo did not account for when the same entity is used as both the 1st/2nd association entry for different data entry.
This resulted in the entity node only having one of the Literature<entity> label, causing a mismatch with what the association
said the entity type is. Added a FOREACH to the cyphers to accomodate for this and a few unit tests to confirm data is as expected.