Exclude AssociationType and Publication from pruning.
The bad data seems to be only for LiteratureEntity type nodes, so exclude those two. Another reason
is the code to create those nodes are in a very old repo that needs to be migrated over to the current repo.
So to save time, these nodes will be kept. If we find bad data in them in the future, we can make a new changeset.