Added validCheckSum tag to the changesets.
This tag is used to bypass calculated checksum that is done by Liquibase. Liquibase calculates checksums based on
the content of a changeset, however when a changeset is changed slightly, the checksum is different.
This causes a validation failure in Liquibase. This is not a good implementation by Liquibase, but it is what it is.
The validCheckSum is currently NOT implemented in the liquibase-neo4j, this PR is to take advantage of it when it is.