Fix image upload behavior bug
Our order of events for pasting content on to the drawing tool was:
image > text. This explains why text pasted from MS Office tools was
being pasted as an image instead. Probably the clipboard saves both
plaintext and image data when copying from those programs, and we just
happened to be preferring images.

Also, the image paste handler was stopping propagation of the paste
event, which is why you also couldn't paste that text in the "Node
Details" field, for example.

*Finally*, we weren't checking if the user was actually pasting into the
canvas when pasting image nodes, so the event was triggered regardless
of where you pasted in the app. For example, you could paste in the
search bar of the file browser, and the image paste handler would pick
it up.