Set user state to use most recent auth token props
Previously, we were using whatever was stored in postgres, which might
be out of sync with what's in the oauth server. We don't (currently)
update both when the user updates their profile, so we have to be
careful to use one or the other. Maybe we should also update postgres
whenever the user updates their profile, although that opens up an
entirely different can of syncing issues.