Fix token refresh flow
There was a property access bug in the failed token refresh flow that
was preventing the app from redirecting to the login screen. Fixed this,
and removed a few unused methods from the oauth service. Also, fixed an
issue where the user was not properly redirected to the initial login
target after signing in.