Update auto refresh to logout user on failure
It seemed odd to me that the user would get sent to the login page
immediately. My experience on most sites is that when your login expires
you don't get redirected to the login page immediately, and can instead
browse while unauthenticated. Commented out the original code in case we
change our minds, but I think this is the more user-friendly approach.