Fix LL-4302
In the `Files` ORM class, we were returning what was essentially an
empty object when there was no fallback organism defined for the file.
This was causing problems on the round trip from the
backend->frontend->backend, because those fields are expected to have
values, where instead they were null.