Fix LL-4285
Just had an incorrect positioning calculation. Was erroneously
offsetting the tooltip position by the distance of the canvas from the
top of the window. We can simply use the absolute y-position of the
hovered text element, minus a small offset to make sure the tooltip
isn't overlapping the hovered text.