Add performance compromises to pass pytests
Had to add the manual `modified_date` update and the re-querying for the
file update response to quickly enable pytests to pass. There are some
fundamental changes that need to be added to our tests to remove these
compromises, but in the interest of time leaving it like this for now.