Update config to ensure `None` values for JWT vars
This hadn't been a problem up to this point because all of our env vars
were strings, or we always set values for the JWT vars. In the future we
may want to update the code to evaluate certain env vars for their true
values, but this works in the meantime.